backend/requirements.txt
fastapi>=0.111.0,<1.0
uvicorn[standard]>=0.30.0,<1.0
supabase>=2.5.0,<3.0
httpx>=0.27.0,<1.0
pydantic>=2.7.0,<3.0
pydantic-settings>=2.3.0,<3.0
redis>=5.0.0,<6.0
PyJWT>=2.8.0,<3.0
